perf: lazy-load GSAP animations
Load GSAP only when the current page has animation hooks and remove the unused HeroSection plugin import.\n\nFixes #12
