+++ src/components/sections/HeroSection.astro
@@ -76,7 +76,7 @@ const t = useTranslations(locale);
 </section>
 
 <script>
-    import { gsap, ScrollTrigger } from "../../lib/gsap";
+    import { gsap } from "../../lib/gsap";
 
     document.addEventListener("astro:page-load", () => {
         const heroBg = document.getElementById("hero-bg");
